cat-sense/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.9.2
python-dotenv>=1.0.0
numpy>=1.26.0
scipy>=1.13.0
snowflake-connector-python>=3.0.0
google-generativeai>=0.8.0
pillow>=10.0.0
anthropic>=0.20.0
openai>=1.0.0
